feat: added archive workspace, invite redeem
