+++ server/src/modules/organization-users/service.ts
@@ -26,6 +26,7 @@ import { IOrganizationUsersService } from './interfaces/IService';
 import { UpdateOrgUserDto } from './dto';
 import { RequestContext } from '@modules/request-context/service';
 import { AUDIT_LOGS_REQUEST_CONTEXT_KEY } from '@modules/app/constants';
+import { Organization } from '@entities/organization.entity';
 @Injectable()
 export class OrganizationUsersService implements IOrganizationUsersService {
   constructor(
@@ -83,22 +84,36 @@ export class OrganizationUsersService implements IOrganizationUsersService {
   }
 
   async archive(id: string, organizationId: string, user?: User): Promise<void> {
-    const organizationUser = await this.organizationUsersRepository.findOneOrFail({
-      where: { id, organizationId },
-      relations: ['user'],
-    });
+    await dbTransactionWrap(async (manager: EntityManager) => {
+      const organizationUser = await manager.findOneOrFail(OrganizationUser, {
+        where: { id, organizationId },
+        relations: ['user'],
+      });
 
-    await this.organizationUsersUtilService.throwErrorIfUserIsLastActiveAdmin(organizationUser?.user, organizationId);
-    await this.organizationUsersRepository.update(id, {
-      status: WORKSPACE_USER_STATUS.ARCHIVED,
-      invitationToken: null,
+      await this.organizationUsersUtilService.throwErrorIfUserIsLastActiveAdmin(organizationUser?.user, organizationId);
+      await manager.update(OrganizationUser, id, {
+        status: WORKSPACE_USER_STATUS.ARCHIVED,
+        invitationToken: null,
+      });
+      await this.setAuditLog(organizationUser, manager);
     });
+  }
 
+  async setAuditLog(organizationUser: OrganizationUser, manager: EntityManager) {
+    console.log('organization user', organizationUser.user);
+    const organization = await manager.findOne(Organization, {
+      where: { id: organizationUser.organizationId },
+    });
+    console.log('organizaiton data', organization);
     const auditLogEntry = {
       userId: organizationUser.userId,
       organizationId: organizationUser.organizationId,
       resourceId: organizationUser.user.id,
       resourceName: organizationUser.user.email,
+      resourceData: {
+        workspaceName: organization.name,
+        workspaceId: organization.id,
+      },
     };
 
     RequestContext.setLocals(AUDIT_LOGS_REQUEST_CONTEXT_KEY, auditLogEntry);
@@ -155,6 +170,7 @@ export class OrganizationUsersService implements IOrganizationUsersService {
 
       await this.licenseUserService.validateUser(manager);
       await this.licenseOrganizationService.validateOrganization(manager);
+      await this.setAuditLog(organizationUser, manager);
     });
 
     if (organizationUser.user.invitationToken) {
@@ -171,14 +187,6 @@ export class OrganizationUsersService implements IOrganizationUsersService {
           sender: user.firstName,
         },
       });
-
-      const auditLogEntry = {
-        userId: organizationUser.userId,
-        organizationId: organizationUser.organizationId,
-        resourceId: organizationUser.user.id,
-        resourceName: organizationUser.user.email,
-      };
-      RequestContext.setLocals(AUDIT_LOGS_REQUEST_CONTEXT_KEY, auditLogEntry);
 
       return;
     }
